+++ src/main/java/com/arcao/wmt/ui/fragment/TrackableDetailFragment.java
@@ -6,7 +6,7 @@ import com.arcao.wmt.data.database.model.AbstractTrackableModel;
  * Created by Arcao on 31. 5. 2014.
  */
 public class TrackableDetailFragment extends AbstractTrackableFragment {
-	public static <M extends AbstractTrackableModel> TrackableDetailFragment newInstance(Class<M> modelClass, int id) {
+	public static <M extends AbstractTrackableModel> TrackableDetailFragment newInstance(Class<M> modelClass, long id) {
 		TrackableDetailFragment fragment = new TrackableDetailFragment();
 		fragment.setArguments(createArguments(modelClass, id));
 		return fragment;
